Fix Highness Heal calculation : Fixes bugreport:7966 Map crash fix when casting Kaihou - Fixes bugreport:7962 Update Songs/Dances for Land Protector - Fixes bugreport:7913 Show baseATK increases in status window - Fixes bugreport:7960 Revert damage to int64 - Fixes bugreport:7939 - Also removed limit on Acid Demonstration Changed Camouflage to increase Equip Attack - Partially fixes bugreport:7941

+++ src/map/battle.h
@@ -46,7 +46,7 @@ struct Damage {
 #ifdef RENEWAL
 	int statusAtk, statusAtk2, weaponAtk, weaponAtk2, equipAtk, equipAtk2, masteryAtk, masteryAtk2;
 #endif
-	int damage,damage2; //right, left dmg
+	int64 damage,damage2; //right, left dmg
 	int type,div_; //chk clif_damage for type @TODO add an enum ? ;  nb of hit
 	int amotion,dmotion;
 	int blewcount; //nb of knockback
